changed ft_strcutuntil

+++ includes/matrix.h
@@ -14,7 +14,7 @@ typedef enum		e_mat_error
 	E_IDX_ZERO,
 }					t_mat_error;
 
-typedef double		t_any;
+typedef int64_t		t_any;
 
 typedef struct		s_matrixdesc
 {
@@ -30,8 +30,8 @@ t_matrix			mat_create_nxm(size_t rows, size_t cols);
 t_matrix			mat_create_4x4(void);
 
 void				mat_resize(t_matrix a, size_t rows, size_t cols);
-void				mat_swap(t_matrix a, t_matrix b); 
-void				mat_free(t_matrix* a);
+void				mat_swap(t_matrix a, t_matrix b);
+void				mat_free(t_matrix *a);
 
 /** Accessors **/
 t_any				mat_get(const t_matrix a, size_t i, size_t j);
@@ -41,7 +41,7 @@ size_t				mat_rows(const t_matrix a);
 size_t				mat_cols(const t_matrix a);
 
 /** Operations **/
-void				mat_mul_mat(t_matrix a, t_matrix b, t_matrix c); /* a = b * c */
+void				mat_mul_mat(t_matrix a, t_matrix b, t_matrix c); /* a = b + c */
 void				mat_mul_scalar(t_matrix a, t_any x);
 void				mat_add(t_matrix a, t_matrix b, t_matrix c); /* a = b + c */
 
@@ -49,4 +49,3 @@ void				mat_add(t_matrix a, t_matrix b, t_matrix c); /* a = b + c */
 void				mat_error_handle(const t_mat_error me, const char *func);
 
 #endif
-
